// src/app/customers/models/customer-put.ts
export class CustomerPut {
  Id: Number;
  Name: String;
  RG: String;
  CPF: String;
  CNPJ: String;

  Email: String;
  HomePhone: String;
  CommercialPhone: String;
  CellPhone: String;

  CEP: String;
  StreetAddress: String;
  Complement: String;
  District: String;
  City: String;
  State: String;
}